TF-IDF Embeddings (#25)
Implements TF-IDF text embeddings. Performs classification on these embeddings using Logistic Regression, but not the other models, due to slow training time.

There are 70K terms / columns in the document term matrix, which makes training very slow. It also makes it hard to save the resulting file of embeddings, which is 5GB as an HD5 file 😮 

So we should consider limiting the number of terms or doing dimensionality reduction or something.